ADD: styles for adaptive project section
