+++ src/components/ProjectSection/ProjectSection.jsx
@@ -1,13 +1,9 @@
-import { useMediaQuery } from "react-responsive";
-import { useScreenWidth } from "../../hooks/useScreenWidth";
 import data from "./data";
 
-import { Title, List, ListItem, HiddenInfo } from "./ProjectSection.styled";
-import DetailInfo from "./UI/DetailInfo/DetailInfo";
-import ListItemMarkup from "./UI/ListItemMarkup/ListItemMarkup";
+import { Title, List } from "./ProjectSection.styled";
+import ProjectSectionCard from "./ProjectSectionCard/ProjectSectionCard";
 
 const ProjectSection = () => {
-  const screenWidth = useScreenWidth();
   return (
     <section
       id="projects"
@@ -17,12 +13,7 @@ const ProjectSection = () => {
       <Title>Проєкти</Title>
       <List>
         {data.map((item) => (
-          <ListItem key={item.category}>
-            <ListItemMarkup dataList={item} screenWidth={screenWidth} />
-            <HiddenInfo image={item.image}>
-              <DetailInfo dataList={item} />
-            </HiddenInfo>
-          </ListItem>
+          <ProjectSectionCard key={item.category} item={item} />
         ))}
       </List>
     </section>
