+++ src/components/TickList/TickList.jsx
@@ -1,9 +1,16 @@
 import icon from "../../images/sprite.svg";
-import { List, ListItem, Title, Icon, Comment } from "./TickList.styled";
+import {
+  List,
+  ListItem,
+  Title,
+  Icon,
+  Comment,
+  Wrapper,
+} from "./TickList.styled";
 
 const TickList = ({ title, about, comment, additionalComment = false }) => {
   return (
-    <>
+    <Wrapper>
       <Title forDetail={true}>{title}</Title>
       <List>
         {about?.map((item) => (
@@ -18,7 +25,7 @@ const TickList = ({ title, about, comment, additionalComment = false }) => {
       {comment && (
         <Comment additionalComment={additionalComment}>{comment}</Comment>
       )}
-    </>
+    </Wrapper>
   );
 };
