+++ src/components/TickList/TickList.styled.js
@@ -1,4 +1,9 @@
 import styled from "@emotion/styled";
+export const Wrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+`;
 
 export const Title = styled.h4`
   font-size: ${(props) => (props.forDetail ? "16px" : "12px")};
