+++ src/index.css
@@ -79,7 +79,7 @@ textarea {
 }
 
 .container {
-  width: 100%;  
+  width: 100%;
   max-width: 375px;
   margin-left: auto;
   margin-right: auto;
